Fix ASM classloader conflicts.
These has a couple pieces. First, the cobertura configuration no longer
extends testCompile. This is necessary because the cobertura files need
to be further isolated from what the project may be importing. Secondly,
we need to exclude the log4j jars from the configuration else we get
ClassLoader issues with the JCL classes. Basically, this amounts to all
JCL/Log4j files being loaded from the parent class loader instead of the
URLClassloader created to handle the cobertura configuration jars.