Add link to our community wiki for Troubleshooting
This should help those users like us at Ericsson that faced numerous issues that have been reported about missing dependencies and general instrumentation errors and where we solved them all after several days and debugging with Gradle and Cobertura.  That information has now been provided in the new Wiki sections I created.  I also mentioned this as a comment in a few issues that were identical to ours and that we solved for the community.  All the best.